Replaced Audio "Performance" Fix with Better One
Okay. So instead of adding randomness to timestep two commits ago, I add
randomness to spawn. I wanted to do this to begin with, but I was
thinking of adding a random delay to life. This might hurt WebCL and
WebGL as life > 1.

Instead, I just ignore a dead particle until it wins an RNG check. This
probability aligns with an expected value of any frame in a "decay"
timespan.