Utilize Local Memory as a Look-up Table
It's a test to see how much random Global Memory access was hurting me
(a lot). Intel HD 4600 is up to ~20,000-30,000 sounds.

Note that it's currently unrealistic, but represents a future
optimization. This only allows ~250-350ms of sound cues... total...
which works for this demo but not much else. This principle could lead
to a two-kernel setup, where the first kernel assembles each output
sample's contibuting samples in order, and the second executes it. This
would be a much nicer access pattern, albeit a bit more overhead than
this.